refactor: move BarGraph to chatbotDashboard shared components
