+++ frontend/src/features/chatbotDashboard/components/shared/BarGrapgh.tsx
